refactoring: merge two packages to simplify
